Redid entirety of news posts
`octopress-quote-tag` is not maintained and is broken with current Jekyll
so I had to find a new way to do the news posts without it. Now uses
`collections` to generate each post, pretty closely mimicking old look,
except the page that lists all news posts has lost the extra formatting.